Hiccup while adding state
adding state to the evaluator is where I'm now stumbling through
the type system, I just haven't developed an intution for it yet,
and I think its primariy because of not paying attention to where
I see certain tokens.

I've deviated from the paper slightly, in that I've followed the runM
patter forn unwrapping my computations.

The next issue though: adding state, comprises a situation where
I now posses Mc a, a type constructor for types supported by runM.

The next issue of course is determining how to make the functions I've
defined so far, play well in this new context.

The author has stressed that the pride and pain of fp is that
data flow is explicit. -- I can't call a function without also supplying
the parameters it requires to operate, as there is no global state the
function can call upon to alter its behaviour.  That is unless the
operation is monadic, in which case the monad, whilst acting as the
container for data, can also at the discretion of how it was designed,
provide.

So I get all that, but what I'm failing to grasp is how to articulate
this computation in a way where it's type valid and mentally coherent.

Composing these nested compuations is where I need the most practise.

Regards,
Dante Elrik
Syntk Games Inc.